refactor: organise components
