+++ app/components/members/InvitedMembersList.tsx
@@ -1,17 +1,83 @@
-import { useLoaderData } from '@remix-run/react'
+import { useLoaderData, useSubmit } from '@remix-run/react'
+import moment from 'moment'
 import { useTranslation } from 'react-i18next'
-import type { Invites } from '~/interface/Interface'
-import InvitedMembersListItems from './InvitedMemberListItem'
+import type { User, Role, Invites } from '~/interface/Interface'
+import Table from '../common-components/TableComponent'
+import memberResendIcon from '~/../public/assets/resend-member-invitation.svg'
 
 const InvitedMembersList = ({
   actionStatus,
+  invitedMemberCurrentPage,
+  setInvitedMemberPage,
+  invitedMemberPageSize,
+  setInvitedMemberPageSize,
 }: {
   actionStatus: string | undefined
+  invitedMemberCurrentPage: number
+  setInvitedMemberPage: (e: number) => void
+  invitedMemberPageSize: number
+  setInvitedMemberPageSize: (e: number) => void
 }) => {
   const membersData = useLoaderData()
   const { t } = useTranslation()
-
   const invitedMember = membersData.invitedMembers
+  const submit = useSubmit()
+  const memberLoaderData = useLoaderData()
+
+  const InvitedByCell = (data: Invites) => {
+    return (
+      <span>
+        {data.invitedById.firstName} {data.invitedById.lastName}
+      </span>
+    )
+  }
+  const InvitedOnCell = (data: Invites) => {
+    return <span>{moment(data?.invitedOn).format('DD MMMM YY')}</span>
+  }
+  const RoleDataCell = (data: { role: Role }) => {
+    return <span>{data.role.name}</span>
+  }
+  const resendMail = (id: string) => {
+    let data = {
+      id: id,
+      action: 'resendMember',
+    }
+    submit(data, {
+      method: 'post',
+    })
+  }
+  const InviteAction = (data: User) => {
+    return (
+      <span
+        tabIndex={0}
+        role="button"
+        onKeyUp={(e) => {
+          if (e.key === 'Enter') resendMail(data.id)
+        }}
+        onClick={() => resendMail(data.id)}
+        className="cursor-pointer opacity-100"
+      >
+        <img src={memberResendIcon} alt="reinvite" id="resend-member-invite" />
+      </span>
+    )
+  }
+  const invitedMembersColumn = [
+    { title: 'Email', field: 'email', width: '30%' },
+    { title: 'Role', field: 'role', render: RoleDataCell },
+    {
+      title: 'Invited By',
+      field: 'invitedById',
+      render: InvitedByCell,
+      width: '25%',
+    },
+    {
+      title: 'Invited On',
+      field: 'invitedOn',
+      width: '20%',
+      render: InvitedOnCell,
+    },
+    { title: 'Action', field: 'action', render: InviteAction },
+  ]
   return (
     <>
       {invitedMember?.length !== 0 && (
@@ -24,35 +90,17 @@ const InvitedMembersList = ({
           >
             {t('members.invitedMember')}
           </h1>
-
-          <div className="grid grid-cols-12 rounded-lg shadow-base">
-            <div className="col-span-full grid grid-cols-10 rounded-lg border border-solid border-gray-200 bg-white">
-              <div className="col-span-full grid grid-cols-10 gap-3 bg-gray-100 py-4 px-6">
-                <h1 className="col-span-3 pl-4 text-sm text-gray-500">
-                  {t('commonConstants.email')}
-                </h1>
-                <h1 className="col-span-2 pl-4 text-sm text-gray-500">
-                  {t('members.role')}
-                </h1>
-                <h1 className="col-span-2 pl-4 text-sm text-gray-500">
-                  {t('resultConstants.invitedBy')}
-                </h1>
-                <h1 className="col-span-2 pl-4 text-sm text-gray-500">
-                  {t('members.invitedOn')}
-                </h1>
-                <h1 className="col-span-1 pl-4 text-sm text-gray-500">
-                  {t('members.action')}
-                </h1>
-              </div>
-              {invitedMember.map((invitedMembers: Invites, index: number) => (
-                <div key={index} className="memberRow col-span-10 grid">
-                  <InvitedMembersListItems
-                    actionStatus={actionStatus}
-                    invitedMembers={invitedMembers}
-                  />
-                </div>
-              ))}
-            </div>
+          <div className="text-base">
+            <Table
+              columns={invitedMembersColumn}
+              data={memberLoaderData.invitedMembers}
+              paginationEnabled={true}
+              pageSize={invitedMemberPageSize}
+              setPageSize={setInvitedMemberPageSize}
+              currentPage={invitedMemberCurrentPage}
+              onPageChange={setInvitedMemberPage}
+              totalItems={memberLoaderData.invitedUsersCount}
+            />
           </div>
         </div>
       )}
